ISSUE #4321 - Fixed issues and risk schema circular dependencies
